Moved `RatedVoteCase` again to voting_algorithms.py
